test: run standalone TLSSocket server-wrap tests on both Node.js and Bun
Move the three scenarios into a runtime-agnostic node:test file
(node-tls-socket-server-wrap.test.mts) that runs under bun bd test and
node --test, with a Bun-side wrapper that spawns node on the same file,
following the node-http2-upgrade.test.mts pattern.
