+++ test/js/node/tls/node-tls-upgrade.test.ts
@@ -2,7 +2,6 @@ import { expect, test } from "bun:test";
 import { once } from "events";
 import { tls as certs } from "harness";
 import net from "net";
-import { Duplex } from "node:stream";
 import tls from "tls";
 
 test("should be able to upgrade a paused socket and also have backpressure on it #15438", async () => {
@@ -60,137 +59,4 @@ test("should be able to upgrade a paused socket and also have backpressure on it
   }
 
   expect().pass();
-});
-
-test("new TLSSocket({ isServer, requestCert }) without ca requests a client certificate", async () => {
-  // https://github.com/oven-sh/bun/issues/33954
-  const { promise: securePromise, resolve: resolveSecure, reject } = Promise.withResolvers<tls.PeerCertificate>();
-  const rawServer = net.createServer(raw => {
-    const socket = new tls.TLSSocket(raw, {
-      isServer: true,
-      key: certs.key,
-      cert: certs.cert,
-      requestCert: true,
-    });
-    socket.on("secure", () => resolveSecure(socket.getPeerCertificate()));
-    socket.on("data", data => socket.write(data));
-    socket.on("error", reject);
-  });
-  rawServer.listen(0, "127.0.0.1");
-  await once(rawServer, "listening");
-  const { port } = rawServer.address() as net.AddressInfo;
-
-  const client = tls.connect({
-    host: "127.0.0.1",
-    port,
-    key: certs.key,
-    cert: certs.cert,
-    rejectUnauthorized: false,
-  });
-  client.on("error", reject);
-  const echoPromise = new Promise<string>(resolveEcho => client.on("data", data => resolveEcho(data.toString())));
-  try {
-    const peerCert = await securePromise;
-    expect(peerCert.subject).toMatchObject({ CN: "server-bun" });
-    // The presented certificate is untrusted, but a standalone server-side
-    // TLSSocket must not auto-reject the connection: Node applies that policy
-    // only in tls.createServer's connection listener.
-    client.write("ping");
-    expect(await echoPromise).toBe("ping");
-  } finally {
-    client.end();
-    rawServer.close();
-  }
-});
-
-test("new TLSSocket({ isServer }) without requestCert does not request a client certificate", async () => {
-  const { promise: securePromise, resolve: resolveSecure, reject } = Promise.withResolvers<tls.PeerCertificate>();
-  const rawServer = net.createServer(raw => {
-    const socket = new tls.TLSSocket(raw, {
-      isServer: true,
-      key: certs.key,
-      cert: certs.cert,
-    });
-    socket.on("secure", () => resolveSecure(socket.getPeerCertificate()));
-    socket.on("error", reject);
-  });
-  rawServer.listen(0, "127.0.0.1");
-  await once(rawServer, "listening");
-  const { port } = rawServer.address() as net.AddressInfo;
-
-  const client = tls.connect({
-    host: "127.0.0.1",
-    port,
-    key: certs.key,
-    cert: certs.cert,
-    rejectUnauthorized: false,
-  });
-  client.on("error", reject);
-  try {
-    const peerCert = await securePromise;
-    expect(Object.keys(peerCert ?? {})).toHaveLength(0);
-  } finally {
-    client.end();
-    rawServer.close();
-  }
-});
-
-test("new TLSSocket(duplex, { isServer, requestCert, rejectUnauthorized }) requests a certificate and does not auto-reject", async () => {
-  // Generic-Duplex variant of the standalone server wrap (no native fd), so it
-  // exercises the upgradeDuplexToTLS path instead of the fd-backed upgrade.
-  const a: Duplex = new Duplex({
-    read() {},
-    write(chunk, _enc, cb) {
-      b.push(chunk);
-      cb();
-    },
-    final(cb) {
-      b.push(null);
-      cb();
-    },
-  });
-  const b: Duplex = new Duplex({
-    read() {},
-    write(chunk, _enc, cb) {
-      a.push(chunk);
-      cb();
-    },
-    final(cb) {
-      a.push(null);
-      cb();
-    },
-  });
-
-  const { promise: securePromise, resolve: resolveSecure, reject } = Promise.withResolvers<tls.PeerCertificate>();
-  const server = new tls.TLSSocket(a, {
-    isServer: true,
-    key: certs.key,
-    cert: certs.cert,
-    requestCert: true,
-    rejectUnauthorized: true,
-  });
-  server.on("secure", () => resolveSecure(server.getPeerCertificate()));
-  server.on("data", data => server.write(data));
-  server.on("error", reject);
-
-  const client = tls.connect({
-    socket: b,
-    key: certs.key,
-    cert: certs.cert,
-    rejectUnauthorized: false,
-  });
-  client.on("error", reject);
-  const echoPromise = new Promise<string>(resolveEcho => client.on("data", data => resolveEcho(data.toString())));
-  try {
-    const peerCert = await securePromise;
-    expect(peerCert.subject).toMatchObject({ CN: "server-bun" });
-    // The cert is untrusted and rejectUnauthorized is set, but a standalone
-    // server-side TLSSocket never auto-rejects: Node applies that policy only
-    // in tls.createServer's connection listener.
-    client.write("ping");
-    expect(await echoPromise).toBe("ping");
-  } finally {
-    client.end();
-    server.destroy();
-  }
 });
